src/requirements.txt
google-cloud-bigquery==3.25.0
requests==2.32.3
pendulum==3.0.0
